refactor(useMedia): update useMedia to include support for MatchMedia
